refactor: ThemeProvider 설정
