+++ front/src/utils/index.tsx
@@ -1,0 +1,3 @@
+export { default as theme } from './theme';
+// export { Outer } from './styled';
+export { default as getDecoPoisition } from './position';
