+++ front/src/utils/position.tsx
@@ -7,28 +7,28 @@ const getDecoPoisition = (n: number): THREE.Vector3 => {
     [0.5, 1.5],
     [-1.5, 0.5],
     [-0.5, -1.5],
-    [1.5, 0.5], //5
+    [1.5, 0.5], // 5
     [-0.5, 1.5],
     [-1.5, -0.5],
     [0.5, -1.5],
     [1.5, -1.5],
-    [1.5, 1.5],
+    [1.5, 1.5], // 10
     [-1.5, 1.5],
     [-1.5, -1.5],
     [2.5, -0.5],
     [0.5, 2.5],
-    [-2.5, 0.5],
+    [-2.5, 0.5], // 15
     [-0.5, -2.5],
     [2.5, 0.5],
     [-0.5, 2.5],
     [-2.5, -0.5],
-    [0.5, -2.5],
+    [0.5, -2.5], // 20
     [2.5, -1.5],
     [1.5, 2.5],
     [-2.5, 1.5],
-    [-1.5, -2.5], //24
-    [2.5, 1.5],
-    [-1.5, 2.5], //26
+    [-1.5, -2.5],
+    [2.5, 1.5], // 25
+    [-1.5, 2.5],
     [-2.5, -1.5],
     [1.5, -2.5],
     [2.5, -2.5],
@@ -37,4 +37,4 @@ const getDecoPoisition = (n: number): THREE.Vector3 => {
   return new THREE.Vector3(positions[n][0], 0, positions[n][1]);
 };
 
-export { getDecoPoisition };
+export default getDecoPoisition;
